ThesisManagementSystem/requirements.txt

django==3.2.15
djangorestframework==3.14.0
numpy==1.21.6
scikit-learn==1.0
PyPDF2==2.11.0
gunicorn==20.1.0
python-decouple==3.6
psycopg2-binary==2.9.3